Update to the latest macaw
The latest macaw changed the representation of the MemSegmentOff
type (specifically, SegmentIndex is now gone, partially replaced by
RegionIndex).  The changes to do a direct port would have been fairly awkward,
so I took the opportunity to simplify our address representation (and to choose
a better name).

The RelAddress type has been replaced by ConcreteAddress.  It is now backed by
the macaw MemAddr type, but is a newtype with the additional restriction that
the MemAddr contained in a ConcreteAddress is always an absolute address (i.e.,
it has a base address of zero).  This greatly simplifies address arithmetic, as
we no longer have to keep segment indexes consistent and normalized.  This
reflects and implicit choice we had made w.r.t. the binaries supported by the
rewriter: we only rewrite a single module at a time, so we never need to address
basic blocks living in other modules.  Those are treated as opaque jumps to code
that we can't address or rewrite.

The implied limitation is that, if we are given a dynamically linked binary to
rewrite, we will only be able to rewrite code in that binary and are not able to
instrument code in called shared libraries.  This seems like a reasonable
restriction -- users needing to also instrument shared libraries should
instrument those libraries separately.

Some of the address operations (especially address difference) in macaw return
maybes for invalid operations.  Because we have a restriction on our newtyped
ConcreteAddress, these operations become total.  The partiality is hidden in the
Renovate.Address module and is safe.