Fix a bug in fallthrough reification during layout
The successor index was only being computed from the modified blocks; if the
successor of a modified block was not modified (and we needed it), the we would
later error out because we had missed a fallthrough successor.

The fix is to compute the fallthrough index based on *all* of the blocks.