Change the analysis callback API
The callback frequency was causing an inordinate amount of work to be redone.
This commit has two changes:

 1) A throttle
 2) A lazier approach

The throttle lets the caller have some control over the frequency with which the
function discovery callback is called.  The lazier approach potentially means
that the entire result is produced lazily, requiring much less effort per call.