Fix a bug in rewriting conditional branches
There is a longer description in Note [Conditional Branch Restrictions].  In
short, conditional branches only have enough offset bits to jump +/- 32KB, which
isn't enough for realistic binaries.  We have to work around this by generating
a more elaborate instruction sequence that uses a short conditional branch to
jump to some unconditional branches (which have 26 bits of offset, which is good
for +/- 32MB).

This change also includes the removal of an unused argument, which would have
been hard to plumb to the spot that needed it in the fix for this bug.