Add support for larger binaries in PowerPC
This was a fairly involved change that uncovered a number of bugs and required a
bit of redesign.  The motivating problem was that the old choices of address for
code layout were too constrained to support realistically-sized programs since
they put instrumented code on top of original code.  Choosing new addresses
posed one major problem that drove the rest of the changes: conditional branches
in PowerPC only have 16 bits of offset, which is not enough to reach the
instrumentation section (or, more importantly, back to the original code *from*
the instrumentation section).

The fix for this involves translating short indirect jumps into a sequence like

```
; Original code
ble <target>

;; New code
ble 8
b <fallthroughTarget>
b <target>
```

Essentially, we transform short conditional jumps into longer unconditional
jumps, which let us cover +/- 32MB of code, which is enough for now.

This required a change to the API of `isaSymbolizeAddresses`, which before could
only return a single instruction for its input.  This change requires being able
to map one original instruction to three new instructions.  This change also
cleaned up some other APIs around `isaModifyJumpTarget`, which was previously
being abused (unsafely in several ways) to promote x86 jumps from three to five
bytes.