x86: Fix a nasty bug in isaSymbolizeAddresses
The wrong instruction was being used as the base for the map of
saveAbsoluteRipAddresses.  It was using the *original* instruction, which can
have the wrong opcode (e.g., for short jumps that have been promoted to long
jumps).  This manifest later as isaModifyJumpTargets appearing to change the
size of a jump.  Really, it was isaSymbolizeAddresses that was allocating a jump
of the wrong size.