Avoid instrumenting blocks in incomplete functions
A function is incomplete if any of its basic blocks ended in a ClassifyFailure.
Instrumenting a function that is incomplete can break control flow (e.g., if we
move a block that really is split by a jump that we didn't discover).  This
approach is much safer (though it still has an assumption about not jumping
between functions without a call).