Add Crucible CFGs for complete discovered functions
This change uses macaw-symbolic to convert recovered macaw functions into
crucible CFGs, where possible.  We currently only convert complete
functions (i.e., functions with no blocks ending in ClassifyFailure or decoding
errors).  This commit only constructs CFGs for x86, as the PowerPC symbolic
backend isn't ready yet.

This introduces a new dependency on macaw-symbolic (and the relevant x86
backend).

There was also a change to SomeConfig to expose the analysis result type - it
turns out that we need are going to need that (otherwise we can't really write
an analysis consumer).