Expose a bit more information from macaw-symbolic to the arch-specific backends
We need to plumb this information through in the same way we did with the
symbolic functions object.  Unfortunately, it has a funny type that prevents us
from returning it directly - MacawArchEvalFn is a typedef around a function with
a forall in it, which is not something we can return without
-XImpredicativePolymorphism, which is pretty broken.

Instead, we have to put it in an argument position, which means a with-style
function.  As an alternative, we could cook up a data wrapper around the
function and return that, but it doesn't really save us anything.