Add closer integration with the BinaryLoader interface
Instead of the analysis engine taking a macaw 'Memory' object, it now takes a
'LoadedBinary', which contains extra information.  For example, on PowerPC it
contains the Table of Contents (TOC), which is very useful during analysis (for
identifying indirect call targets).

Most of the complexity is around invoking the 'BinaryLoader' interface in the
Elf loading code.  Ideally, we could simplify all of that code at some point in
the future.  The problem right now is that we need to parse the elf file to
determine the architecture of the binary, and we can't invoke the 'BinaryLoader'
interface until we know both the binary format (ELF) and the architecture.  This
means we have to have the ELF file pre-parsed before we can learn arch-specific
information.

Someday in the future, we will ideally just have a function that takes a
bytestring and does all of the necessary dispatching to produce a loaded binary.