Substantially simplify binary loading
This relies on some changes to macaw-semmc that introduce a general binary
loading interface that preserves arch-specific metadata.  This allows us to
simplify the configuration of the PowerPC loaders, as they can just use this
arch-specific data after the loading is completed.

This approach seems to help client code avoid re-inspecting ELF files multiple
times.