Make `RewriterT` resumable
As an alternative to making `RewriteM` a transformer to get access to
`IO` in analysis.

Analysis used to completely independent of rewriting, but now that
analysis gets access to the symbolic version of the program, it needs
to come after the symbolic blocks are computed, which happens in the
`RewriterT` monad. The actual rewriting happens inside the nesting of
the `RewriteM` and `RewriterT` monads, and so the choices for getting
`IO` in analysis are either to make `RewriteM` into a transformer or
to make it possible to run analysis outside of those monads, by making
`RewriterT` resumable. Here I went with making `RewriterT` resumable,
because I was curious to see how that would turn out.