Revise the ELF rewriting strategy, add a new tool (refurbish)
The old rewriting strategy ran into some problems with some of the latest
elf-edit updates.  The old strategy was also not correct for binaries that used
TLS (and libc supporting it).  The problem mostly came down to the kernel
assuming that the program header table (PHDRs) was always in the first loadable
segment.

The new strategy currently works for both x86 and PowerPC and is roughly as
follows:

 * Wipe out all of the old variable length sections (symbol table, section
   table, program header table) and replace them with padding to preserve the
   original binary layout
 * Append the new text section as a new segment
 * Put a fresh PHDR table at the end of the binary in its own loadable
   segment *but* with a segment index of 0 (suitably updating all of the other
   segment indexes) so that it comes first in the PHDR table.  This meets the
   kernel requirement while giving us the flexibility to add new segments.

Note that we also force the address of the PHDR table to be very high in the
address space, as the kernel also requires that the address of the PHDR table be
at least as large as its offset into the file.

This commit also added a new tool, refurbish, to test the rewriter.  This is
handy, as it can pull in the arch-specific backends without putting any
arch-specific dependencies on the core renovate library.  The tool allows for
identity rewriting of PowerPC and x86 ELF binaries.