Expand the test suite and fix a few bugs in the layout (mostly affecting PowerPC)
The new test suite includes PowerPC and x86 binaries compiled against glibc as a
stress test.  Not all of the tests currently pass, but they are something to
work towards.

The fixes cover two major classes of (related) problem:
 1. Relocation/symbolization was failing with an error (sometimes)
 2. Layout around overlapping blocks was broken

Fixing the first problem made the second immediately apparent.

The relocation code was previously fragile in that, if renovate did not find a
block that corresponded to the target of a jump, it would call error.  This is
problematic for a number of reasons, but is in general not a good behavior.

Problem 1: The macaw heuristic for identifying functions only called
indirectly (i.e., by function pointer) behaves badly on PowerPC.  The heuristic
says that any pointer stored to memory that points into an executable segment
should be considered to be a function pointer.  On PowerPC, the Table of
Contents (TOC) is mapped into executable memory, which means that macaw tries to
treat it as code.  This isn't correct, and produces lots of invalid
code (sometimes ending in jumps to nothing sensible).  It also produced many
overlapping blocks, which cause their own failures.

Fix 1: Restrict renovate to only generate blocks for code in the .text section.
This throws away the "blocks" that macaw finds that actually correspond to the
TOC.

Problem 2: Restricting renovate to the .text section means that it never finds
valid code that really does live in other sections (e.g., .init and .fini).
This caused the symbolization to fail aggressively, because there were jumps to
valid code for which renovate had no block.

Fix 2: Support jumps to concrete addresses for which we have no target block.
This was an extension to the SymbolicAddress type.  The original SymbolicAddress
is used to represent an address reference from a program location that can move
to another program location that could also move.  The change adds a new type of
symbolic address that occurs at a program location that can move, but references
a program location that cannot move (e.g., a function in .init, which we don't
rewrite).  This makes the whole process a lot more robust.

Problem 3: Handling of overlapping blocks was incorrect, and caused errors
during analysis and while re-assembling.

Fix 3: Add some new machinery (Renovate.Recovery.Overlap) for computing which
blocks overlap other blocks and prevent callers from making modifications to any
overlapping blocks.  The data structure is based on an IntervalMap and makes it
efficient to find overlapping blocks.  The instrumentor is never called on
blocks that overlap other blocks.  This required some updates in the basic
block assembly/text section reconstruction code.  That code attempted to permit
limited rewriting of overlapping blocks, but was incorrect.  This functionality
isn't particularly necessary (for now at least), so simplify everything by
simply not supporting rewrites of overlapping blocks.

A few minor things that ended up changing as a result:
 * SymbolicAddress is now parameterized by arch (since it can now contain a
   concrete address)
 * Some diagnostics have additional information