Fix a bug in basic block recovery
The recovery function has a check to reject basic blocks discovered outside of
the expected address range of code.  The original check was failing more often
than intended because it was using `concreteAsSegmentOff` to convert
`ConcreteAddress` to `MemSegmentOff` (which was then converted to `MemWord`).
The conversion to `MemSegmentOff` failed because we were computing the address
one past the end of the acceptable address range, which fails because
`concreteAsSegmentOff` cannot create "invalid" addresses.

The fix is to convert directly to `MemWord` using `absoluteAddress`, which
cannot fail.