Fix another rewriter error
The problem was a stray check in the block assembly code.  This check came from
the time where we were trying to support rewriting cleverly-overlapped blocks.
The code wasn't right anyway, and we don't support that anymore, so remove the
check.

I think the original check was trying to assert that, if a set of blocks
overlapped, the smaller blocks contained within the largest formed a
non-overlapping tiling.  The original goal was that we could add redirections
for each of the contained overlapping blocks (as long as space permitted), but
there isn't a good reason to support that (and it can't always work out anyway).