rework the way redirected blocks get truncated
Previously, when we redirected a basic block to a new location, we
replaced it with a suitable branch followed by some ISA-specific padding
to ensure that the resulting block was still the same size. This is fine
for the parallel layout engine, but leads to shenanigans with the
compact layout: in the compact layout, we plan to reuse space we freed
up by redirecting that block, so this introduces a spurious conflict
between padding from a redirected block and an actual instruction stream
from a moved (and maybe rewritten) block.

The new plan goes like this: when redirecting, we just return a shorter
block if we're able to. Then it is the layout computation's
responsibility to fill any space it reclaims but doesn't use with
ISA-specific padding. For the compact layout engine, this means putting
padding in any spaces left over in the heap; for the parallel layout
engine it means putting padding in all reclaimed spaces.

This way we don't produce conflicting overlapping blocks, so the
assembler doesn't have to do anything complicated to choose between
alternative byte streams.