Fix a codegen bug (observed on PowerPC, but probably widely applicable)
The observed behavior was a segfault in rewritten binaries.  Tracing indicated
that a jump taken at run-time led to a jump target that was not consistent with
renovate's view of the block structure of the program.

It turned out that the jump in question was never updated during
concretization.  Tracing that further indicated that the jump had in fact never
even been assigned a sybmolic target.  The case deleted by this commit was the
culprit, which prevented that symbolic target from being allocated.

It isn't clear what this code was originally trying to do.  I suspect that it
was trying to accommodate jumps within a block (though the check was incorrect
for that case, anyway).  Another possibility is that it was trying to not adjust
jumps outside of the text section, which is also not actually covered by what
was being checked.  Deleting it seems to be the right thing to do.