Add a new pre-analysis callback to renovate (for AnalyzeAndRewrite)
This is useful to allocate e.g., new global memory before analysis, in cases
where analysis results need to refer to fresh globals.  This required being able
to run IO in RewriterM, which was a small but unfortunate change.  It isn't a
huge problem, however; callers can't see the difference.

Note: the documentation examples are out of date again and need to be updated