choose extratext location dynamically
When rewriting a binary, we now search for a suitable virtual address to
load the extratext section's segment. Previously it was placed at a
fixed address; this caused problems for binaries that already had
segments there. The new procedure is to find the largest chunk of unused
addresses that are within easy branch distance of the existing text
section's segment.