Enhance the refurbish test suite to also run rewritten binaries
The output from the unmodified binaries is compared against the modified
binaries.  Details are in the README file in the refurbish/tests directory.  The
enhanced harness uses QEMU (in user mode) and Docker (for isolation).  If Docker
is not available, the tests still rewrite binaries, but do not execute them.  A
message is printed in that case.