WIP More debugging
The new strategy is to expand the old data section to cover the BSS (with actual
zeros).  This rewriting seems to work.  The blocker is now that elf-edit seems
to exhibit a disagreement between `elfLayoutSize layout` and
`B.length (elfLayoutBytes layout)`.  This turns into an off-by-4 error in the
final assembly.  Interestingly, sometimes the elf-edit internal consistency
checks catch this, but other times they miss it (and produce an invalid binary).