Fix .bss initialization in rewritten binaries
The observed bug was that the bss section in binaries was not being initialized,
despite being mapped.  It turns out that this is due to some choices in the
Linux kernel ELF loader.  It maintains a high-watermark of the highest loaded
segment address seen so far.  It only initializes .bss if the last loaded segment
has a smaller physical size than memory size and is less than the highest
address seen.  We map the program headers at a very high address as the first
segment (to work around the kernel not mapping them properly if they are not in
the first segment); this breaks .bss initialization.

The fix is to remove the .bss section and expand the .data section to cover the
address range of the .bss, explicitly initialized with zeros.  There is a bit of
extra work required to maintain the alignment of the data that comes after the
data section in the binary.

Detailed comments are in the new Renovate.BinaryFormat.ELF.BSS module.