Fix the code injection API
During concretization, the symbolic address->concrete address mapping for
injected code was not combined with the mapping for other blocks, leading to
failures later in the process (symbolic addresses were not found when looked up
to be concretized).

This change includes a (small) test suite for the code injection API.  The base
binary fails with an exit(1), while code injection is used to call an injected
function that instead calls exit(0).