Fix another injected code handling bug
Injected functions were captured before redirection.  This is incorrect, as the
rewriting function itself can add new functions (though it isn't especially
likely).  The fix is to require that the base monad for redirection be able to
provide the most up-to-date view of the injected functions.

In practice, this probably means that only a single base monad is possible, but
it isn't restricted yet.