Fix the code injection API
During concretization, the symbolic address->concrete address mapping for
injected code was not combined with the mapping for other blocks, leading to
failures later in the process (symbolic addresses were not found when looked up
to be concretized).

This change includes a (small) test suite for the code injection API.  The base
binary fails with an exit(1), while code injection is used to call an injected
function that instead calls exit(0).

Fix another injected code handling bug where injected functions were captured
before redirection.  This is incorrect, as the rewriting function itself can add
new functions (though it isn't especially likely).  The fix is to require that
the base monad for redirection be able to provide the most up-to-date view of
the injected functions.

In practice, this probably means that only a single base monad is possible, but
it isn't restricted yet.

Also allow refurbish tests to be run in parallel.