Pipe Dockerfile into Docker rather than as command-line argument
This has the (rather surprising) effect of causing Docker to run with an
empty context rather than using the current directory, which saves
copying the whole context over to the daemon.  This isn't so bad here
(the Refurbish root is only ~6MB), but it was taking a bunch of time on
startup in Brittle (whose root is over 4.5GB on my machine!).