Fix support for adding data to binaries during rewriting
Add a new properly-aligned section after the new text section.  The address is
stable and does not need correction (to take into account alignment), as it
starts out properly aligned.

The section is currently limited to a page.

We add a new data section unconditionally (it just contains a page of zeros if
the caller did not request anything) to support running binaries under
qemu-user.  qemu-user ELF loader differs from the kernel loader in that it sets
the program break (brk) after the last data segment, rather than after the last
mapped segment.  Without a trailing data segment, this causes problems with our
new text section (the first time the program calls brk).