branch rewriting megacommit
This begins a reworking of the way branches are rewritten. Previously
renovate was in charge of inserting unconditional branches to redirect
control flow. This was unfortunate because renovate was forced to commit
up front to adding branches where none might eventually be needed to
accomodate the worst case where the targets of the branches were very
far away. A common case is for the targets to be very close, sometimes
even the very next instruction.

This commit begins reworking the rewriting logic to delay this decision
until the targets are known. We still reserve enough space to do the
worst-case situation of two far jumps (one for the target and one for
the fallthrough), but allow the ISA developer to code up special cases
when shorter jump sequences are possible.