WIP Work on making x86 failures cleaner
There are some unhandled cases for glibc-based static binaries from macaw and
flexdis.

The flexdis failures occur when we can disassemble an instruction, but can't
re-assemble it.  The candidate fix is to refuse to build renovate blocks for
code we can't re-assemble.

Currently looking at some of the macaw-derived failures.  It seems possible that
it could be invalid instructions that are causing doJump and getCallTarget in
macaw-x86 to fail.  Either those cases need to be handled, or failures in those
functions need to be made more robust (or both).