Modify an assertion to actually pass
This assertion is trying to verify that, if one block overlaps another, one is a
suffix of the other.  We are trying to ensure that, if blocks overlap, they end
at the same place.

The old check was at the instruction level, but that isn't good enough if the
blocks overlap at the instruction level (i.e., one instruction is a superset of
the bytes of another).  This is very uncommon, but is actually used in x86 with
the `lock` prefix on some instructions.  When not necessary, some code will jump
into the middle of an instruction to skip the lock prefix.

The new check is at the level of bytes to accommodate this.