Fix a layout bug
In rare case, a block could be rewritten such that, if it had fall-through
control flow, the fallthrough successor would not be respected (leading to
crashes or worse).

In particular, it would happen if the actual fallthrough successor for a piece
of code could not be disassembled into renovate's IR for whatever reason.  The
index that renovate used to compute successors is not correct in the presence of
gaps, which made the layout choose the wrong fallthrough successor.

The temporary fix is to just check to see if the chosen successor is really
directly adjacent to the current block - if it isn't, that means that there is a
gap that renovate can't handle, and that the fallthrough should be to the
original machine code.  We force this with a StableAddress successor.

The long-term fix will be to fix the index construction.