Rewrite the symbolic successor computation
This map is used to determine the fallthrough control flow successor for blocks during layout, and
is critical for correctly stitching control flow back together during layout as we are rearranging blocks.  The
old method for computing it involved zipping blocks with their immediate successors.  This was
incorrect in cases where renovate skipped a region of code due to decoding errors.  We had added a
hack to make it work in those cases (a defensive check based on the address of the block flagged as
the successor).

This commit replaces the hack with a more direct indexing method.