Tweak the extra code layout option to be easier to use
It is difficult for callers (especially from the command line) to figure out what is safe to pass
in, as the code requires the offset + original address + text size to be aligned.  This change
modifies the layout code to treat the user input as advisory (e.g., an order of magnitude offset)
and forcibly aligns the requested address to something safe.