relocate unmodified blocks
Previously we tried to avoid adding fallthrough information to
unmodified blocks. Since unmodified blocks were always left where they
are, it would be strange to try to rewrite their jumps to change their
fallthrough behavior, when there might not even be space available to
include the new fallthrough instructions. Anyway this was just a safety
net, as we arranged never to pass unmodified blocks to the functions
that add fallthrough information anyway.

Now that we keep loop bodies and function bodies together, though, we
don't always leave unmodified blocks in place. Sometimes they need to be
relocated to keep them together with modified blocks. So it is no longer
correct to have that safety net in place; if I tell you to add
fallthrough information to a block, you hecking well have to do it for
the resulting rewritten program to be correct.