Introduce CI-checkable examples embedded in documentation
This change uses doctest to check that usage examples in haddock still compile in the face of changes (assuming that the test suite is run in CI).  The haddocks don't render quite as well as before, as each definition has to be in its own GHCi multiline input block.  I think the tradeoff is worth it, though.  This change moves the haddock examples of using renovate into Refurbish.Tutorial (from Renovate.Tutorial) - this is needed because the examples depend on renovate-x86 and renovate-ppc.

The old examples directory is removed in favor of this code that is actually checked.