fix the refurbish tests
We now track the difference between "we decided this block is
modifiable, but the analysis phase chose not to rewrite it" and "we
decided this block is not modifiable". The code for doing compact layout
previously assumed all Unmodified blocks corresponded to the former,
leading to bugs where it attempted to lay out code over blocks that we
must not redirect.