don't lie about PHDRs
We have to write new program headers when we create a rewritten ELF.
Since we don't want the contents of the ELF to shift around just because
our new headers are bigger, we overwrite the old headers with zero and
add a new segment for the new headers. But this is dangerous, because
now we might have a PHDR segment with wrong information in it (all
zeros!); and in particular this causes glibc to do odd and incorrect
things.

So now we additionally change any existing PHDR segments into NULL ones.