Split ELF rewriting state into a r/w and read-only environment
Specifically, move the calculation of the base address for the new text section
into a pre-computing phase, lifting it into a more generic monad. It's good to
be able to know where the new text section will go before actually rewriting
the binary. My plan is to use this downstream in SFE.

Also, this removes the need to uphold an invariant about computing this address
before beginning to rewrite the binary; the calculation is no longer performed
in the rewriting monad.