Replace the fully-generic BasicBlock type with specialized types (#28)
This commit revises renovate core to use concrete block types instead of type aliases over a shared `BasicBlock` type.  The overly-polymorphic type made it difficult to introduce some important things (e.g., attaching a macaw `ParsedBlock` to the `ConcreteBlock`).  Additionally, the typing led to confusing errors and made it difficult to understand the evolution of data in the core.  This commit also adds some more descriptive names to the intermediate rewriting steps.  It removes some partiality in the code by forcing basic blocks to be empty (this was already a requirement, but now it is enforced at the type level).

This change is a useful prerequisite to the next change, which will move the computation of control flow fallthroughs earlier in the rewriting process and changing how they are represented.  Currently, control flow fallthrough addresses are annotations on instructions.  They will be changed into annotations on blocks, which is now possible now that there is no longer a shared type (i.e., there is a place to put them).

Mostly fixes #26.