Further revise jump target rewriting
This commit adds some type safety to force us to only call `isaModifyJumpTarget`
on instructions that are really jumps (that function now requires type-level
evidence that is true).

This actually forced a restructuring of the size computation and concretization
code, which made both of those pieces of code simpler.

The motivation for this change was to simplify the architecture-specific
implementations of `isaModifyJumpTarget`, which now no longer needs to do quite
as much error checking (since the new target address is no longer optional).