Prevent a panic in `isaMakeRelativeJumpTo`
We were passing in a fake address that was actually out of range of jumps,
causing a panic.  Since we always use jumps of the same size, it doesn't matter
what fake address we pass in as long as it doesn't cause the panic.  Just pass
the same address in as the source and target in the fake address generation cases.