Fix a layout bug affecting PowerPC
The bug popped up in the layout configurations that kept loop bodies and
function bodies contiguous.  The problem arose due to a disparity in code
sequence sizes when a short (conditional) jump was sufficient to reach the
target and a long jump was not required.  This imbalanced the block size
caused the rest of the layout code to fill the extra space with halts to pad out
the block size.

Unfortunately, since this changeset separated the insertion of fallthrough jumps
from jump target rewriting, this meant that the rewritten short conditional jump
fell through into halt instructions before hitting the actual fallthrough jump
inserted later, crashing the program.

This fix inserts the necessary pair of no-ops to preserve the size of the block
and prevent this crash.