Relax an overly-aggressive consistency check
The failure case was triggered if we ever encountered a jump instruction with a
rewritable target, but did not have a target to rewrite to.  It turns out that
this is a useful capability in rewriter clients to insert code in basic blocks
that is locally skipped.  This commit simply relaxes that check to allow the
more flexible behavior.

Note: it would be nice to force client code to somehow mark this type of
inserted jump explicitly to ensure that we don't introduce silent errors.