Fix some address calculations
This includes a fix for a case where conditional branch terminators were not
being correctly recognized, and were thus never rewritten.

The address calculations have to be offset by 8 due to an oddity in the way
reading the PC works in ARM (reading the PC yields PC+8 due to prefetch).