Get data reference rewriting working in AArch32
This only works for non-PIE binaries for now.

The major change is that isaModifyJumpTarget and isaConcretizeAddreses are now
merged.  They both ultimately resolve symbolic references, but one was
originally for data references while the other was for pc-relative control
references.

In ARM, we need to be able to add extra instructions to rewrite data accesses,
which required allowing isaConcretizeAddresses to return multiple instructions.
This made it difficult to correctly call isaModifyJumpTarget afterwards.  To
simplify the core, this commit combines the two.  For all of our current
architectures, there is no overlap between the two cases anyway, so the
implementations of the resulting combined function are not more complex than the
separate versions.

The details of the necessary code sequence are in the code with comments, but at
a high level the problem is that PC-relative data references can only reach
4kb.  To work around this, we embed the full pointer in the instruction
stream (and ensure that it doesn't get executed), and use a short PC-relative
load to load the real pointer.  This adds an extra indirection, which is not
ideal, but difficult to work around.  It would require additional analysis to
remove the extra indirection.  It should be possible in many cases, though.

This commit also includes a fix for PowerPC, which began triggering a new error
in the test suite (apparently due to some other changes?).  I believe that the
error was that the maximum jump range computation was not taking sign into
account and therefore allowed the new text section to be allocated too far away
to reach with a long jump.  The range has been shortened to fix the issue.  I
suspect that it worked by accident before.