Fix test compilation
An earlier commit removed the unix package, which looked unused, to enable
Windows builds. It turns out that the test suite (which I thought I was
building, but was in fact not) actually uses it to make generated files
executable.

This commit uses the unix-compat package to achieve the same functionality, but
allow portability to Win32.