Tr/warning cleanup (#56)
Submodule updates

Includes warning cleanup and ghc-8.10 compat. CI is currently failing because it depends on docker, which you can't run inside another docker container (i.e., as a github action). The biggest changes were in renovate-ppc, which had to become more systematic in order to be able to implement `isaInstructionRepr`.