Attempt to get the tests working in CI (#58)
We can't use the Docker runner, since Actions is running in docker. This change adds an optional mode that just doesn't use the docker runner (which we use in the CI harness).  The commands are otherwise the same, so it isn't substantially different from what is run locally on a developer machine.  The configuration changes required a bit more work to install the necessary packages (e.g., qemu), which were not required in the original CI setup.  The path mapping inside of the docker runner is now simplified, which also makes this change easier.

Fixes #57