Adapt to the latest elf-edit API changes
The ELF Layout type was removed, which the ELF rewriter in renovate used.  The
replacement is explicit re-rendering of the current ELF into a
bytestring (followed by re-parsing the headers).  This change required some
pre-indexing of section and segment offsets and sizes.

Client code is slightly simplified due to the changed API (there is no longer
duplication between 32 and 64 bit architectures).

As part of this change, a logger is explicitly threaded through the rewriter
and analysis.