Fix the PHDR address search
This search attempts to find an address to put the PHDRs at that is both unused
and that satisfies all of the constraints required to work with both the normal
system loader and QEMU(-user).  See the various comments for a detailed
discussion of those constraints.

The address search code broke because the new method for determining segment
offsets (which this algorithm requires) involves appending a placeholder segment
header table to the ELF file.  The original code assigned this placeholder
segment a very high address, but that then appeared in the list of candidate
addresses and caused the address search into the high end of the address space,
which made the constraints for qemu-user compatibility unsatisfiable.

The fix was to just use 0 as the initial guess address for the fake segment
header table address.

The change uses a more descriptive return value for the `findSpaceForPHDRs`
function to reflect the distinctions between different types of valid address,
which also means that we don't fall back to a fixed PHDR table address (and use
the best we can find).  If we cannot find any valid address, we do just fail
with an exception.